refactor(FoodList.js): moved food list rendering into its own component(RenderFoodLists)
